net: openvswitch: Remove check for tc_recirc_sharing_support
Remove check for OVS support for TC recirc_id sharing, and just assume
it is supported. This enables older OVS versions to use the recirc_id
sharing for ST CT without adding a request for this feature, as they
already support this sharing. Since this kernel is only targeting OVS
buids which support TC recirc_id sharing, this check is unnecessary.

Change-Id: Iefd1236f11b924e61b170ad28bde7e9ea89da3a9
Signed-off-by: Gavi Teitz <gavi@nvidia.com>